backend/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
google-generativeai
moviepy
imageio-ffmpeg
srt
python-dotenv
